Extract miscellaneous request handlers
